Version 10.3.5

+++ Src/OSF/CAIRO-1160/PIXMAN/pixman-gradient-walker.c
@@ -165,10 +165,10 @@ uint32_t _pixman_gradient_walker_pixel(pixman_gradient_walker_t * walker, pixman
 	r = a * (walker->r_s * y + walker->r_b);
 	g = a * (walker->g_s * y + walker->g_b);
 	b = a * (walker->b_s * y + walker->b_b);
-	a8 = a + 0.5f;
-	r8 = r + 0.5f;
-	g8 = g + 0.5f;
-	b8 = b + 0.5f;
+	a8 = static_cast<uint8_t>(a + 0.5f);
+	r8 = static_cast<uint8_t>(r + 0.5f);
+	g8 = static_cast<uint8_t>(g + 0.5f);
+	b8 = static_cast<uint8_t>(b + 0.5f);
 	v = ((a8 << 24) & 0xff000000) | ((r8 << 16) & 0x00ff0000) | ((g8 <<  8) & 0x0000ff00) | ((b8 >>  0) & 0x000000ff);
 	return v;
 }
